Fix norm-series swapping columns
Also tests

+++ ipsiblings/model/timestampseries.py
@@ -11,7 +11,7 @@ class TimestampSeries:
     """
     KEY_TS_VAL = 'ts_val'
     KEY_RECEPTION_TIME = 'reception_time'
-    DTYPE = numpy.dtype([(KEY_TS_VAL, numpy.int32), (KEY_RECEPTION_TIME, numpy.float64)])
+    DTYPE = numpy.dtype([(KEY_TS_VAL, numpy.uint64), (KEY_RECEPTION_TIME, numpy.float64)])
 
     def __init__(self, key: Tuple[int, str, int], data: numpy.ndarray):
         self.ip_version, self.target_ip, self.target_port = key
